# backend/weoucbookcycle_go/docs/swagger.yaml
definitions:
...
  models.Book:
    properties:
      author:
        type: string
      category:
        type: string
      condition:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        description: 存储JSON数组
        type: string
      isbn:
        type: string
      like_count:
        type: integer
      listings:
        items:
          $ref: '#/definitions/models.Listing'
        type: array
      price:
        type: number
      seller:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      seller_id:
        type: string
      status:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  models.Chat:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_message:
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      updated_at:
        type: string
      users:
        description: 关联关系
        items:
          $ref: '#/definitions/models.ChatUser'
        type: array
    type: object
  models.ChatResponse:
    properties:
      id:
        type: string
      last_message:
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      unread_count:
        description: 添加未读数字段
        type: integer
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.ChatUser'
        type: array
    type: object
  models.ChatUser:
    properties:
      chat:
        allOf:
        - $ref: '#/definitions/models.Chat'
        description: 关联关系
      chat_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      unread_count:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.Favorite:
    properties:
      created_at:
        type: string
      id:
        type: string
      listing:
        $ref: '#/definitions/models.Listing'
      listing_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      user_id:
        type: string
    type: object
  models.Listing:
    properties:
      book:
        allOf:
        - $ref: '#/definitions/models.Book'
        description: 关联关系
      book_id:
        type: string
      buyer:
        $ref: '#/definitions/models.User'
      buyer_id:
        type: string
      created_at:
        type: string
      favorite_count:
        type: integer
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      note:
        type: string
      price:
        type: number
      seller:
        $ref: '#/definitions/models.User'
      seller_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.Message:
    properties:
      chat:
        allOf:
        - $ref: '#/definitions/models.Chat'
        description: 关联关系
      chat_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      sender:
        $ref: '#/definitions/models.User'
      sender_id:
        type: string
    type: object
  models.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      books:
        description: 关联关系
        items:
          $ref: '#/definitions/models.Book'
        type: array
      chat_users:
        items:
          $ref: '#/definitions/models.ChatUser'
        type: array
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      last_login:
        type: string
      listings:
        items:
          $ref: '#/definitions/models.Listing'
        type: array
      login_count:
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      phone:
        type: string
      role:
        description: 角色管理 (简单实现：admin, user)
        type: string
      status:
        type: integer
      trustScore:
        description: TrustScore 用户的信任分（0-100）
        type: integer
      updated_at:
        type: string
      username:
        type: string
      verified_at:
        type: string
      wechat_openid:
        description: 微信开放平台openid，用于小程序登录
        type: string
      wishlist_items:
        description: Wishlist 关联改为独立表
        items:
          $ref: '#/definitions/models.Wishlist'
        type: array
    type: object
  models.Wishlist:
    properties:
      book:
        allOf:
        - $ref: '#/definitions/models.Book'
        description: 关联
      book_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/chats:
    get:
      consumes:
      - application/json
      description: 获取当前用户的聊天列表，包含每个聊天的未读消息数
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ChatResponse'
            type: array
      security:
      - Bearer: []
      summary: 获取聊天列表
      tags:
      - chats
  /api/v1/chats/:id:
    delete:
      consumes:
      - application/json
      description: 删除指定聊天
      parameters:
      - description: 聊天ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除聊天
      tags:
      - chats
  /api/v1/chats/:id/read:
    put:
      consumes:
      - application/json
      description: 标记指定聊天的所有未读消息为已读
      parameters:
      - description: 聊天ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 标记消息为已读
      tags:
      - chats
  /api/v1/chats/{id}:
    get:
      consumes:
      - application/json
      description: 根据聊天ID获取详细信息
      parameters:
      - description: 聊天ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Chat'
      security:
      - Bearer: []
      summary: 获取聊天详情
      tags:
      - chats
  /api/v1/chats/{id}/messages:
    get:
      consumes:
      - application/json
      description: 获取聊天的消息列表（分页）
      parameters:
      - description: 聊天ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取聊天消息
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: 在指定聊天中发送新消息
      parameters:
      - description: 聊天ID
        in: path
        name: id
        required: true
        type: string
      - description: 消息内容
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - Bearer: []
      summary: 发送消息
      tags:
      - chats
  /api/v1/chats/online-users:
    get:
      consumes:
      - application/json
      description: 获取当前在线的用户列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取在线用户列表
      tags:
      - chats
  /api/v1/chats/unread:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有未读消息数量
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取未读消息数
      tags:
      - chats
...
swagger: "2.0"